[fix] #3473: Make `kagami validator` runnable from outside `iroha` dir
