+++ wasm_builder/Cargo.toml
@@ -12,4 +12,5 @@ categories = ["development-tools::build-utils"]
 eyre = "0.6.8"
 serde_json = "1.0.99"
 sha256 = "1.1.4"
+path-absolutize = "3.1.0"
 wasm-opt = "0.113.0"
